requirements.txt
# Web UI
gradio>=5.0.0

# Core ML / Deep Learning
# GPU build (CUDA 12.1) — remove the extra-index-url line to fall back to CPU-only
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0

# Hugging Face ecosystem
transformers>=4.35.0
huggingface_hub>=0.25.1
timm>=0.9.0

# Computer Vision
Pillow>=10.0.0
opencv-python>=4.8.0

# audioop backport for Python 3.13 (pydub dependency of gradio 4.x)
audioop-lts; python_version >= "3.13"
fpdf2>=2.7          # PDF report generation

# Data / ML utilities
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.11.0

# Notebooks
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Signature preprocessing (sigver-compatible)
scikit-image>=0.21.0

# OCR with native Italian support
easyocr>=1.7.0

# Date parsing (multilingual, used in Datazione Documenti tab)
dateparser>=1.1

# RAG / knowledge base (Consulente Forense IA tab)
pypdf>=3.0          # text extraction from digital PDFs
pymupdf>=1.23       # render scanned PDF pages as images (for OCR fallback)
python-docx>=0.8    # read DOCX files

# Agentic framework (Agente Documentale tab)
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# Document layout detection — PaddleOCR (Agente Documentale)
# CPU build: paddlepaddle; GPU build: paddlepaddle-gpu
paddlepaddle>=2.6.0
paddleocr>=2.8.0
